pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "treehouse-lab"
version = "1.0.1"
description = "Karpathy-style autoresearch loops for tabular ML"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Matthew Davis" }
]
keywords = ["xgboost", "tabular-ml", "automl", "mlops", "optuna"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "xgboost>=2.1",
  "optuna>=4.0",
  "mlflow>=2.20",
  "pyyaml>=6.0"
]

[project.urls]
Homepage = "https://github.com/CommandWM/treehouse-lab"
Repository = "https://github.com/CommandWM/treehouse-lab"
Issues = "https://github.com/CommandWM/treehouse-lab/issues"

[project.scripts]
treehouse-lab = "treehouse_lab.cli:main"
treehouse-lab-api = "treehouse_lab.api:main"

[project.optional-dependencies]
ui = [
  "streamlit>=1.49"
]
web = [
  "fastapi>=0.115",
  "uvicorn>=0.34"
]
llm = [
  "openai>=1.0",
  "requests>=2.32"
]
featuregen = [
  "openfe>=0.0.12",
  "featuretools>=1.31"
]
dev = [
  "httpx>=0.28",
  "pytest>=8.3",
  "ruff>=0.8"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
